refactor(fragment): rename data_fragment
